+++ pkgs/cupertino_http/CHANGELOG.md
@@ -1,7 +1,11 @@
+## 0.0.5
+
+* Add the ability to set network service type.
+* Add the ability to control multipath TCP connections.
+
 ## 0.0.4
 
 * Add the ability to control caching policy.
-* Add the ability to control multipath TCP connections.
 
 ## 0.0.3